requirements.txt
fastapi
uvicorn
openai
python-dotenv
langdetect
pydub
requests
